Removing local_settings.py by renaming.
local_settings.py should contain deployment specific configuration
which is not suitable for storing in the repository.  It is best
practice not to commit this file as it will conflict with each
development environment.

Moving this to local_settings.py.sample allows us to store it as an example.
However, you would normally put the database config in this file (which this
sample lacks).